requirements.txt
discord.py[voice]>=2.2.2
yt-dlp==2023.3.4
openai==0.27.8